pyproject.toml
[project]
name = "llm-readme-tester"
version = "1.1.0"
description = "LLM-powered unit tests for your README.md"
authors = [{ name = "Atick Faisal", email = "atick@example.com" }]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }

dependencies = [
    "google-genai>=1.10.0",
    "pytest>=8.3.5",
]

[dependency-groups]
dev = [
    "ruff>=0.11.4",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["F", "E", "W", "I", "B", "C4", "SIM", "UP", "N"]
fixable = ["ALL"]


